backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
langgraph>=0.2
langchain-core>=0.3
langchain-anthropic>=0.2
langchain-openai>=0.2
langchain-google-genai>=2.0
httpx>=0.27
python-dotenv>=1.0
pydantic>=2.8
sse-starlette>=2.1
